refactor: Switch to JS (#180)
* refactor: Switch to JS

* ci: No more build script

* ci: Add type-checking step

* chore: Patch broken types in '@rollup/pluginutils'
