+++ .github/workflows/main.yml
@@ -18,5 +18,5 @@ jobs:
           cache: "npm"
           cache-dependency-path: "**/package-lock.json"
       - run: npm ci
-      - run: npm run build
       - run: npm test
+      - run: npm run test:types
