+++ src/utils.js
@@ -1,0 +1,8 @@
+/**
+ * Allows to ignore query parameters, as in Vue SFC virtual modules.
+ *
+ * @param {string} url
+ */
+export function parseId(url) {
+	return { id: url.split("?", 2)[0] };
+}
