--- tsconfig.json
@@ -1,13 +1,0 @@
-{
-	"compilerOptions": {
-		"target": "ES2019",
-		"strict": true,
-		"esModuleInterop": true,
-		"module": "node20",
-		"moduleResolution": "node16",
-		"declaration": true,
-		"outDir": "dist/",
-		"skipLibCheck": true
-	},
-	"files": ["./src/index.ts"]
-}
